+++ Source/Bumpy/VersionFunctions.cs
@@ -8,7 +8,7 @@ namespace Bumpy
         private const string VersionGroupName = "version";
         private const string NumbersGroupName = "numbers";
         private const string LabelGroupName = "label";
-        private const string TagGroupName = "tag";
+        private const string MarkerGroupName = "marker";
 
         private static readonly Regex BumpyRegex = new Regex(@"^(?<numbers>\d+((\.|,)\d+)*)(?<label>[_\-\+\.0-9a-zA-Z]*)$", RegexOptions.Singleline);
 
@@ -87,14 +87,14 @@ namespace Bumpy
             return new BumpyVersion(formattedNumbers, labelGroup.Value, numberDelimiter);
         }
 
-        public static bool TryParseVersionInText(string text, string regexPattern, out BumpyVersion version, out string tag)
+        public static bool TryParseVersionInText(string text, string regexPattern, out BumpyVersion version, out string marker)
         {
             version = null;
 
             var regex = new Regex(regexPattern, RegexOptions.Singleline);
             var match = regex.Match(text);
             var versionGroup = match.Groups[VersionGroupName];
-            var tagGroup = match.Groups[TagGroupName];
+            var markerGroup = match.Groups[MarkerGroupName];
             var success = versionGroup.Success;
 
             if (success)
@@ -102,7 +102,7 @@ namespace Bumpy
                 version = ParseVersion(versionGroup.Value);
             }
 
-            tag = tagGroup.Success ? tagGroup.Value : string.Empty;
+            marker = markerGroup.Success ? markerGroup.Value : string.Empty;
 
             return success;
         }